requirements.txt
streamlit
openai
langchain-openai
langchain-core
langchain-community
python-dotenv
